Adds Selector.Direct and renames Selector to Selector.Options
The uncontrolled keeps same behaviour
