+++ uncontrolled/Selector.jsx
@@ -1,9 +1,9 @@
 import compose from '../lib/compose'
 import statefulValue from '../lib/decorators/statefulValue'
 import statefulOptionFocus from '../lib/decorators/statefulOptionFocus'
-import Selector from '../Selector'
+import {Options} from '../Selector'
 
-const UncontrolledSelector = compose(statefulOptionFocus, statefulValue)(Selector)
+const UncontrolledSelector = compose(statefulOptionFocus, statefulValue)(Options)
 
 UncontrolledSelector.displayName = 'UncontrolledSelector'
